pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "query-forge-pipeline"
version = "0.1.0"
description = "Offline research pipeline scaffold for Query Forge"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "PyYAML>=6.0.3",
    "psycopg[binary]>=3.2.10",
    "requests>=2.32.5",
    "testcontainers[postgres]>=4.8.2",
]

[project.scripts]
query-forge-pipeline = "cli:main"

[tool.setuptools]
py-modules = ["cli"]

[tool.setuptools.packages.find]
where = ["."]
include = [
    "collectors*",
    "preprocess*",
    "loaders*",
    "common*",
    "generation*",
    "gating*",
    "memory*",
    "datasets*",
    "eval*",
    "tests*",
]
